Rename package multiconsumer to multiprocessor
